+++ Tests/SwiftiomaticTests/Layout/ImportTests.swift
@@ -36,7 +36,6 @@ struct ImportTests: LayoutTesting {
       import class MyModule.MyClass
       import struct MyModule.MyStruct
       @testable import testModule
-
       @_spi(STP) @testable import testModule
 
       """